requirements.txt
flask
python-dotenv
requests
google-generativeai
gspread
oauth2client
pytz
